% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildURL.R
\name{buildURL}
\alias{buildURL}
\title{Build an ERDDAP URL}
\usage{
buildURL(dataStruct, lonBounds, latBounds, timeBounds,
  urlbase = "https://coastwatch.pfeg.noaa.gov/erddap/griddap/")
}
\arguments{
\item{dataStruct}{A structure describing the dataset from erddapStruct.rda}

\item{lonBounds}{A list or reals size two of the longitude bounds}

\item{latBounds}{A list of reals size 2 of the latitude bounds}

\item{timeBounds}{A list of strings size two of the time bounds}

\item{urlbase}{A character string giving the base URL of the ERDDAP server}
}
\value{
The ERDDAP URL to call
}
\description{
Build an ERDDAP URL
}
\keyword{ERDDAP}
\keyword{URL}
\keyword{\code{buildURL}}
\keyword{an}
\keyword{builds}
\keyword{from}
\keyword{function}
\keyword{information}
\keyword{internal}
\keyword{is}
\keyword{provided}
\keyword{that}
\keyword{the}

