% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkBounds.R
\name{checkBounds}
\alias{checkBounds}
\title{Check Coordinate Dimensions}
\usage{
checkBounds(dataStruct, xposLim, yposLim, tposLim)
}
\arguments{
\item{dataStruct}{A structure describing the dataset from erddapStruct.rda}

\item{xposLim}{A list of reals size 2 that contains the longitude limits of
the data request}

\item{yposLim}{A list of reals size 2 that contains the latitude limits of
the data request}

\item{xposLim}{A list of strings  size 2 that contains the time limits of the
data request}
}
\value{
returnCode of 0 if all data is in bounds of the served data, -1
  otherwise
}
\description{
Check Coordinate Dimensions
}
\keyword{\code{checkBounds}}
\keyword{an}
\keyword{are}
\keyword{bounds}
\keyword{checks}
\keyword{dataset}
\keyword{function}
\keyword{internal}
\keyword{is}
\keyword{latitude,}
\keyword{longitude,}
\keyword{requested}
\keyword{th}
\keyword{that}
\keyword{the}
\keyword{time}
\keyword{within}

