% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/colors.R
\docType{data}
\name{colors}
\alias{colors}
\title{cmocean colors
The cmocean color palette by Kristen Thyng as of late 2017}
\format{An object of class \code{list} of length 18.}
\usage{
colors
}
\description{
str(colors)
List of 18
$ algae
$ amp
$ balance
$ curl
$ deep
$ delta
$ dense
$ gray
$ haline
$ ice
$ matter
$ oxy
$ phase
$ solar
$ speed
$ tempo
$ thermal
$ turbid
}
\keyword{datasets}
