% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getErddapURL.R
\name{getErddapURL}
\alias{getErddapURL}
\title{Download Data using an ERDDAP URL}
\usage{
getErddapURL(myURL, destfile, verbose = FALSE)
}
\arguments{
\item{myURL}{A character string containing the full ERDDAP URL}

\item{destfile}{A character string with the name of the file to store the
returned data, a netcdf file}
}
\value{
downloadReturn returns 0 is successful, -1 otherwise
}
\description{
Download Data using an ERDDAP URL
}
\keyword{ERDDAP}
\keyword{URL}
\keyword{\code{getErddapURL}}
\keyword{an}
\keyword{function}
\keyword{internal}
\keyword{is}
\keyword{makes}
\keyword{request}
\keyword{the}
\keyword{to}

