% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getETOPO.R
\name{getETOPO}
\alias{getETOPO}
\title{retrieve ETOPO Bathymetry Data in a Bounding Box}
\usage{
getETOPO(dataStruct, xpos1, ypos, verbose,
  urlbase = "http://coastwatch.pfeg.noaa.gov/erddap/griddap/")
}
\arguments{
\item{dataStruct}{A structure describing the dataset from erddapStruct.rda}

\item{xpos1}{A list of reals size 2 of the longitude bounds}

\item{ypos}{A list of reals size 2 of the latitude bounds}

\item{verbose}{Logical variable if true will produce verbose
   output from httr:GET}

\item{urlbase}{A character string giving the base URL of the ERDDAP server}
}
\value{
Named Data array with data, or else NaN
}
\description{
retrieve ETOPO Bathymetry Data in a Bounding Box
}
\keyword{Bathymetry}
\keyword{Bounding}
\keyword{Box}
\keyword{Data}
\keyword{ETOPO}
\keyword{Function}
\keyword{\code{getETOPO}}
\keyword{a}
\keyword{an}
\keyword{by}
\keyword{given}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{retrieve}
\keyword{xpos,}
\keyword{ypos}

