% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getETOPOtrack.R
\name{getETOPOtrack}
\alias{getETOPOtrack}
\title{retrieve ETOPO Bathymetry Data along a track}
\usage{
getETOPOtrack(dataStruct, xpos1, ypos, xrad, yrad, verbose,
  urlbase = "http://coastwatch.pfeg.noaa.gov/erddap/griddap/")
}
\arguments{
\item{dataStruct}{A structure describing the dataset from erddapStruct.rda}

\item{xpos1}{A list of reals of the track longitudes}

\item{ypos}{A list of reals of the track latitude bounds}

\item{xrad}{A list of reals with distance around given longitude
  to make extract}

\item{yrad}{A list of reals with distance around given latitude
  to make extract}

\item{verbose}{Logical variable if true will produce verbose
   output from httr:GET}

\item{urlbase}{A character string giving the base URL of the ERDDAP server}
}
\value{
Named Data array with data, or else NaN
}
\description{
retrieve ETOPO Bathymetry Data along a track
}
\keyword{Bathymetry}
\keyword{Bounding}
\keyword{Box}
\keyword{Data}
\keyword{ETOPO}
\keyword{Function}
\keyword{\code{getETOPOtrack}}
\keyword{a}
\keyword{an}
\keyword{by}
\keyword{given}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{retrieve}
\keyword{xpos,}
\keyword{ypos}

