% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getMaxTime.R
\name{getMaxTime}
\alias{getMaxTime}
\title{Extract dataset information for a given dtype name or number}
\usage{
getMaxTime(dataStruct,
  urlbase1 = "http://coastwatch.pfeg.noaa.gov/erddap/tabledap/allDatasets.csv?")
}
\arguments{
\item{dataStruct}{A structure describing the dataset from erddapStruct.rda}

\item{urlbase}{A character string giving the base URL of the ERDDAP server}
}
\value{
dataStruct with dataStruct$maxTime updated
}
\description{
Extract dataset information for a given dtype name or number
}
\keyword{\code{getMaxTime}}
\keyword{dataset}
\keyword{for}
\keyword{given}
\keyword{internal}
\keyword{latest}
\keyword{retrieves}
\keyword{the}
\keyword{time}

