% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make180.R
\name{make180}
\alias{make180}
\title{Convert Longitudes to (-180,180)}
\usage{
make180(lon)
}
\arguments{
\item{lon}{A vector of longitudes}
}
\value{
A vector of longitudes all mapped to (-180,180)
}
\description{
Convert Longitudes to (-180,180)
}
\keyword{(-180,180)}
\keyword{(0,360)}
\keyword{\code{make180}}
\keyword{a}
\keyword{an}
\keyword{converts}
\keyword{from}
\keyword{function}
\keyword{internal}
\keyword{is}
\keyword{longitudes}
\keyword{of}
\keyword{that}
\keyword{to}
\keyword{vector}
