% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make360.R
\name{make360}
\alias{make360}
\title{Convert Longitudes to (0,360)}
\usage{
make360(lon)
}
\arguments{
\item{lon}{A vector of longitudes}
}
\value{
A vector of longitudes all mapped to (0,360)
}
\description{
Convert Longitudes to (0,360)
}
\keyword{(-180,180)}
\keyword{(0,360)}
\keyword{\code{make360}}
\keyword{a}
\keyword{an}
\keyword{converts}
\keyword{from}
\keyword{function}
\keyword{internal}
\keyword{is}
\keyword{longitudes}
\keyword{of}
\keyword{that}
\keyword{to}
\keyword{vector}

