% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Marlintag38606.R
\docType{data}
\name{Marlintag38606}
\alias{Marlintag38606}
\title{Marlin Tag Data}
\format{A data frame with 152 obs. of 7 variables:
\describe{
  \item{date}{time of observation given as YYYY-MM-DD}
  \item{lon}{longitude of observation}
  \item{lat}{latitude of observation}
  \item{lowLon}{low error on longitude}
  \item{higLon}{high error on longitude}
  \item{lowLat}{low error on latitude}
  \item{higLat}{high error on latitude}
}}
\source{
Dr. Mike Musyl, Pelagic Research Group LLC
}
\usage{
Marlintag38606
}
\description{
Telemetry data of a blue marlin tagged in the Pacific Ocean in 2003
}
\keyword{datasets}
