% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mbnms.R
\docType{data}
\name{mbnms}
\alias{mbnms}
\title{MBNMS Boundaries}
\format{A data frame with 6666 obs. of 2 variables:
\describe{
  \item{Longitude}{Longitude of a boundary polygon}
  \item{Latitudes}{Latitudes of a boundary polygon}
}}
\source{
\url{http://sanctuaries.noaa.gov/library/imast_gis.html}
}
\usage{
mbnms
}
\description{
A dataset containing the latitudes and longitudes of a polygon that define
 boundaries of the Monterey Bay National Marine Sanctuary.
}
\keyword{datasets}

