% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotBBox.R
\name{plotBBox}
\alias{plotBBox}
\title{plot result of xtracto_3D or rxtracto_3D}
\usage{
plotBBox(resp, plotColor = "viridis", time = NA, animate = FALSE,
  name = NA, myFunc = NA, maxpixels = 10000)
}
\arguments{
\item{resp}{data frame returned from xtracto()}

\item{plotColor}{the color to use in plot from rerddap color palette}

\item{time}{a function to map multi-time to one,  or else identity for animation}

\item{animate}{if multiple times, if TRUE will animate the maps}

\item{name}{name for colorbar label}

\item{myFunc}{function of one argument to transform the data}

\item{maxpixels}{maximum numbe rof pixels to use in making the map - controls resolution}
}
\value{
a plotdap plot
}
\description{
\code{plotBox} is a function to plot the results from
rxtracto() and xtracto()
}
\examples{
require("plotdap")
xcoord <- c(230, 245)
ycoord <- c(30, 45)
tcoord <- c('2006-01-15', '2006-01-18')
extract <- rxtracto_3D('erdMBsstd1day', xcoord, ycoord, tpos = tcoord)
plotBBox(extract, plotColor = 'temperature')
}
