% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotTrack.R
\name{plotTrack}
\alias{plotTrack}
\title{plot result of xtracto or rxtracto}
\usage{
plotTrack(resp, xcoord, ycoord, plotColor = "viridis", name = NA,
  myFunc = NA, shape = 20, size = 0.5)
}
\arguments{
\item{resp}{data frame returned from rxtracto() or xtracto()}

\item{xcoord}{passed to rxtracto() or xtracto()}

\item{ycoord}{passed to rxtracto() or xtracto()}

\item{plotColor}{the color to use in plot from rerddap}

\item{name}{name for colorbar label}

\item{myFunc}{function of one argument to transform the data}

\item{shape}{shape to use to mark track}

\item{size}{size of shape to use to mark track}
}
\value{
a plotdap plot
}
\description{
\code{plotTrack} is a function to plot the results from
rxtracto() and xtracto()
}
\examples{
tagData <- Marlintag38606
xpos <- tagData$lon
ypos <- tagData$lat
tpos <- tagData$date
urlbase <- 'http://upwell.pfeg.noaa.gov/erddap'
swchlInfo <- rerddap::info('erdSWchla8day')
swchl <- rxtracto('erdSWchla8day', xpos, ypos, tpos = tpos, xlen = .2, ylen = .2)
plotTrack(swchl, xpos, ypos,  plotColor = 'chlorophyll')
}
