% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getETOPO.R
\name{getETOPO}
\alias{getETOPO}
\title{retrieve ETOPO Bathymetry Data in a Bounding Box}
\usage{
getETOPO(dataStruct, xpos1, ypos, verbose,
  urlbase = "http://coastwatch.pfeg.noaa.gov/erddap/griddap/")
}
\description{
retrieve ETOPO Bathymetry Data in a Bounding Box
}
\keyword{-}
\keyword{2}
\keyword{@param}
\keyword{@return}
\keyword{A}
\keyword{Bathymetry}
\keyword{Bounding}
\keyword{Box}
\keyword{Data}
\keyword{ERDDAP}
\keyword{ETOPO}
\keyword{Function}
\keyword{Logical}
\keyword{NaN}
\keyword{Named}
\keyword{URL}
\keyword{\code{getETOPO}}
\keyword{a}
\keyword{an}
\keyword{array}
\keyword{base}
\keyword{bounds}
\keyword{by}
\keyword{character}
\keyword{data,}
\keyword{dataStruct}
\keyword{dataset}
\keyword{describing}
\keyword{else}
\keyword{erddapStruct.rda}
\keyword{from}
\keyword{given}
\keyword{giving}
\keyword{httr:GET}
\keyword{if}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{latitude}
\keyword{list}
\keyword{longitude}
\keyword{of}
\keyword{or}
\keyword{output}
\keyword{produce}
\keyword{reals}
\keyword{retrieve}
\keyword{server}
\keyword{size}
\keyword{string}
\keyword{structure}
\keyword{the}
\keyword{true}
\keyword{urlbase}
\keyword{variable}
\keyword{verbose}
\keyword{will}
\keyword{with}
\keyword{xpos,}
\keyword{xpos1}
\keyword{ypos}

