% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getETOPOtrack.R
\name{getETOPOtrack}
\alias{getETOPOtrack}
\title{retrieve ETOPO Bathymetry Data along a track}
\usage{
getETOPOtrack(dataStruct, xpos, ypos, xrad, yrad, verbose,
  urlbase = "https://coastwatch.pfeg.noaa.gov/erddap/griddap/")
}
\description{
retrieve ETOPO Bathymetry Data along a track
}
\keyword{-}
\keyword{@param}
\keyword{@return}
\keyword{A}
\keyword{Bathymetry}
\keyword{Bounding}
\keyword{Box}
\keyword{Data}
\keyword{ERDDAP}
\keyword{ETOPO}
\keyword{Function}
\keyword{Logical}
\keyword{NaN}
\keyword{Named}
\keyword{URL}
\keyword{\code{getETOPOtrack}}
\keyword{a}
\keyword{an}
\keyword{around}
\keyword{array}
\keyword{base}
\keyword{bounds}
\keyword{by}
\keyword{character}
\keyword{data,}
\keyword{dataStruct}
\keyword{dataset}
\keyword{describing}
\keyword{distance}
\keyword{else}
\keyword{erddapStruct.rda}
\keyword{extract}
\keyword{from}
\keyword{given}
\keyword{giving}
\keyword{httr:GET}
\keyword{if}
\keyword{in}
\keyword{internal}
\keyword{is}
\keyword{latitude}
\keyword{list}
\keyword{longitude}
\keyword{longitudes}
\keyword{make}
\keyword{of}
\keyword{or}
\keyword{output}
\keyword{produce}
\keyword{reals}
\keyword{retrieve}
\keyword{server}
\keyword{string}
\keyword{structure}
\keyword{the}
\keyword{to}
\keyword{track}
\keyword{true}
\keyword{urlbase}
\keyword{variable}
\keyword{verbose}
\keyword{will}
\keyword{with}
\keyword{xpos}
\keyword{xpos,}
\keyword{xrad}
\keyword{ypos}
\keyword{yrad}
