% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMaxTime.R
\name{getMaxTime}
\alias{getMaxTime}
\title{Extract dataset information for a given dtype name or number}
\usage{
getMaxTime(dataStruct,
  urlbase1 = "http://coastwatch.pfeg.noaa.gov/erddap/tabledap/allDatasets.csv?")
}
\description{
Extract dataset information for a given dtype name or number
}
\keyword{@param}
\keyword{@return}
\keyword{A}
\keyword{ERDDAP}
\keyword{URL}
\keyword{\code{getMaxTime}}
\keyword{base}
\keyword{character}
\keyword{dataStruct}
\keyword{dataStruct$maxTime}
\keyword{dataset}
\keyword{describing}
\keyword{erddapStruct.rda}
\keyword{for}
\keyword{from}
\keyword{given}
\keyword{giving}
\keyword{internal}
\keyword{latest}
\keyword{of}
\keyword{retrieves}
\keyword{server}
\keyword{string}
\keyword{structure}
\keyword{the}
\keyword{time}
\keyword{updated}
\keyword{urlbase}
\keyword{with}

